Added note reading
Player can pick up note, examine it, use mouse to move it around to explore details(there might be some useful information at the back of the note).

When picking up and reading the note, the note will disappear(like you're actually holding the note and reading it, you cannot open inventory while doing this)

In case the font is difficult to read, we added the read button which provides a more clear version to read.

Fixed the bug that keep pressing e would stack up the note while in examination(whick makes the note edges blur, and game might lag.)